Updated app UI
