--- kute-android-app/.idea/.name
@@ -1,1 +1,0 @@
-kute-android-app
