--- kute-android-app/app/src/main/java/com/kute/app/Activities/SignInActivity.java
@@ -1,23 +1,0 @@
-package com.kute.app.Activities;
-
-
-import android.app.Activity;
-import android.os.Bundle;
-
-import com.kute.app.R;
-
-public class SignInActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.signin_activity);
-
-       /* ImageView splashlogoview=(ImageView)findViewById(R.id.myImage);
-        RelativeLayout mainrelative=(RelativeLayout)findViewById(R.id.splash_mainlayout);
-        splashlogoview.setX(mainrelative.getX()/2);
-        splashlogoview.setY(mainrelative.getY()/2);*/
-
-
-    }
-}
