--- kute-android-app/app/src/main/res/drawable/circle1.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-
-    <solid
-        android:color="#FFF37A6E"/>
-
-    <size
-        android:width="120dp"
-        android:height="120dp"/>
-</shape>
