--- kute-android-app/app/src/main/res/drawable/circle2.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-
-    <solid
-        android:color="#FF9E9D8E"/>
-
-    <size
-        android:width="120dp"
-        android:height="120dp"/>
-</shape>
