--- kute-android-app/app/src/main/res/drawable/roundbutton2.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <corners
-        android:radius="7dp" />
-
-    <solid
-        android:color="#e1c58b"/>
-
-    <size
-        android:width="120dp"
-        android:height="120dp"/>
-
-</shape>
