--- kute-android-app/app/src/main/res/drawable/rounded3.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <corners
-        android:radius="10dp" />
-
-    <solid
-        android:color="#ef9a9a"/>
-
-    <size
-        android:width="120dp"
-        android:height="120dp"/>
-
-</shape>
