+++ kute-android-app/app/src/main/res/layout/activity_map.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true"
-    tools:context="com.kute.app.Activities.Views.MapActivity">
+    tools:context="com.kute.app.Activities.MapActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -22,12 +22,19 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_map" />
+    <include layout="@layout/content_map"/>
 
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@drawable/background"
-        android:alpha="0.2" />
+    <include layout="@layout/bottom_sheet"/>
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/bottom_sheet_fab"
+        app:layout_anchor="@+id/bottom_sheet"
+        app:layout_anchorGravity="top|end"
+        android:src="@drawable/icon_edit_location"
+        android:layout_width="wrap_content"
+        android:layout_marginRight="@dimen/fab_margin"
+        android:layout_marginEnd="@dimen/fab_margin"
+        android:layout_height="wrap_content"/>
+
 
 </android.support.design.widget.CoordinatorLayout>
