+++ kute-android-app/app/src/main/res/layout/content_map.xml
@@ -1,84 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.kute.app.Activities.Views.MapActivity"
-    tools:showIn="@layout/activity_map">
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:orientation="vertical"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent">
 
-    <ImageButton
-        android:id="@+id/train_button"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_gravity="bottom"
-        android:layout_margin="10dp"
-        android:background="@drawable/circle1"
-        android:src="@drawable/icon_train" />
-
-    <ImageButton
-        android:id="@+id/bus_button"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="bottom"
-        android:layout_margin="10dp"
-        android:background="@drawable/circle2"
-        android:src="@drawable/icon_bus" />
-
-    <ImageButton
-        android:id="@+id/car_button"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_gravity="bottom"
-        android:layout_margin="10dp"
-        android:background="@drawable/circle3"
-        android:src="@drawable/icon_car" />
-
-    <LinearLayout
+    <fragment
+        android:id="@+id/map"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginTop="10dp"
-        android:orientation="horizontal"
-        android:layout_above="@+id/train_button">
+        android:layout_height="match_parent"/>
 
-        <Button
-            android:id="@+id/share_button"
-            style="?attr/borderlessButtonStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_margin="8dp"
-            android:layout_weight="1"
-            android:background="@drawable/rounded1"
-            android:text="@string/share_location"
-            android:textSize="20sp" />
-
-        <Button
-            android:id="@+id/show_button"
-            style="?attr/borderlessButtonStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_margin="8dp"
-            android:layout_weight="1"
-            android:background="@drawable/rounded2"
-            android:text="@string/show_location"
-            android:textSize="20sp" />
-    </LinearLayout>
-
-</RelativeLayout>
+</LinearLayout>
