--- kute-android-app/app/src/main/res/layout/content_share_location.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.kute.app.Activities.Views.ShareLocationActivity"
-    tools:shareIn="@layout/activity_share_location">
-
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="@string/share_location_header"
-        android:id="@+id/textView"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:typeface="normal"
-        android:textStyle="bold"
-        android:textSize="25sp"
-        android:layout_marginBottom="10dp" />
-
-    <Spinner
-        android:layout_width="fill_parent"
-        android:layout_height="60dp"
-        android:id="@+id/itemListSpinner"
-        android:layout_below="@+id/textView"
-        android:layout_centerHorizontal="true"
-        android:layout_margin="15dp"
-        android:spinnerMode="dialog"
-        android:background="@drawable/rounded3" />
-
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@+id/itemListSpinner">
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_margin="10dp"
-            android:layout_height="50dp"
-            android:text="@string/share"
-            android:textColor="#000000"
-            android:background="@drawable/roundbutton2"
-            android:id="@+id/share_button"
-            android:layout_weight="1"
-            android:layout_gravity="bottom" />
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_margin="10dp"
-            android:layout_height="50dp"
-            android:background="@drawable/roundbutton1"
-            android:textColor="#000000"
-            android:text="@string/cancel"
-            android:id="@+id/cancel_button"
-            android:layout_weight="1"
-            android:layout_gravity="bottom" />
-    </LinearLayout>
-</RelativeLayout>
