--- kute-android-app/app/src/main/res/layout/layout_individual_share_location.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.kute.app.Views.IndividualShareLocationActivity">
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:alpha="0.2"
-        android:scaleType="fitXY"
-        android:src="@drawable/background"/>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="vertical">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="0dp"
-            android:layout_weight="0.5"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="Train A"
-            android:id="@+id/vehicleID"
-            android:background="@android:color/transparent"
-            android:layout_gravity="center_vertical|center_horizontal"
-            android:gravity="center"/>
-
-        <fragment
-            android:id="@+id/map"
-            android:name="com.google.android.gms.maps.SupportMapFragment"
-            android:layout_width="match_parent"
-            tools:context=".Views.IndividualShareLocationActivity"
-            android:layout_height="0dp"
-            android:layout_marginStart="5dp"
-            android:layout_marginEnd="5dp"
-            android:layout_weight="6"/>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-
-            <Button
-                android:layout_width="0dp"
-                android:layout_weight="1"
-                android:layout_height="50dp"
-                android:layout_margin="10dp"
-                android:text="Search"
-                android:id="@+id/searchStations"
-                android:layout_gravity="center_vertical"
-                android:background="@drawable/roundbutton2"/>
-
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_margin="10dp"
-                android:layout_height="40dp"
-                android:background="@drawable/roundbutton1"
-                android:textColor="#000000"
-                android:text="@string/cancel"
-                android:id="@+id/cancel_button"
-                android:layout_gravity="bottom" />
-
-        </LinearLayout>
-    </LinearLayout>
-</FrameLayout>
