+++ kute-android-app/app/src/main/res/menu/menu_map.xml
@@ -5,12 +5,4 @@
         android:icon="@android:drawable/btn_dialog"
         android:title="@string/sign_out"/>
 
-    <item android:id="@+id/track"
-        android:icon="@android:drawable/btn_dialog"
-        android:title="@string/sign_out"/>
-
-    <item android:id="@+id/clear"
-        android:icon="@android:drawable/btn_dialog"
-        android:title="@string/sign_out"/>
-
 </menu>
