+++ kute-android-app/app/src/main/res/values/colors.xml
@@ -3,4 +3,7 @@
     <color name="colorPrimary">#3F51B5</color>
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
+    <color name="colorCar">#FFF9A825</color>
+    <color name="colorBus">#FF9E9D8E</color>
+    <color name="colorTrain">#FFF37A6E</color>
 </resources>
